Improve AWS related classes: - Support for custom endpoints for all AWS profiles
